fix: about section is component based
